requirements.txt
flask==3.0.3
flask-cors==4.0.0
flask-limiter==3.5.0
requests==2.31.0
twilio==9.0.4
openai==1.30.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pandas==2.0.3
sqlalchemy==2.0.23
